now producing intake catalogue from scratch, re-organised mopdb code #158
