+++ src/mopdata/intake_cat_template.yaml
@@ -1,7 +1,7 @@
 metadata:
   version: 1
 sources:
-  experiment: 
+  <experiment>: 
     description: "Intake catalogue to load ACCESS <experiment> model output"
     Project: ""
     Maintained By: ""
@@ -10,10 +10,11 @@ sources:
     License: "https://creativecommons.org/licenses/by/4.0/"
     Citation: ""
     References: ""
-    driver: intake_esm.esm_datastore
+    driver: intake_esm.core.esm_datastore
     args:
-      obj: "{{CATALOG_DIR}}/catalogue.json"
       columns_with_iterables:
         - variable
         - map_var
         - standard_name
+      read_csv_kwargs: {"dtype": {"date": str}}
+      obj: "{{CATALOG_DIR}}/intake_<experiment>.json"
